requirements.txt
langchain>=0.2.11,<0.3.0
langchain-openai>=0.1.14,<0.2.0
langchain-core>=0.2.23,<0.3.0
langchain-community>=0.2.11,<0.3.0
python-telegram-bot>=21.2
discord.py>=2.3.2
requests>=2.32.3
pyyaml>=6.0.1
google-analytics-admin>=0.25.0
google-analytics-data>=0.18.7
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.127.0
